app/requirements.txt
fastapi
uvicorn
marker-pdf
